fix float --> int for np.ceil()

+++ kabuki/analyze.py
@@ -578,7 +578,7 @@ def plot_posterior_predictive(model, plot_func=None, required_method='pdf', colu
             if not hasattr(bottom_node['node'], required_method):
                 continue # skip nodes that do not define the required_method
 
-            ax = fig.add_subplot(np.ceil(nrows), columns, subj_i+1)
+            ax = fig.add_subplot(np.ceil(nrows).astype(int), columns, subj_i+1)
             if 'subj_idx' in bottom_node:
                 ax.set_title(str(bottom_node['subj_idx']))
